% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaGLMtree.R
\encoding{UTF-8}
\name{deltaGLMtree}
\alias{deltaGLMtree}
\title{Operating area stratification with a delta-GLM}
\usage{
deltaGLMtree(
  binom.model,
  posi.model,
  criteria = "AIC",
  max.narea = 100,
  trend = FALSE,
  delta = 10,
  factor = c("Year", "Area"),
  numeric = NULL,
  seq.length = 10,
  offset = NULL,
  area.w = TRUE,
  lon.name = "Lon",
  lat.name = "Lat",
  min.lon = 140,
  max.lon = 185,
  min.lat = 30,
  max.lat = 50,
  seq.lon = 2.5,
  seq.lat = 2.5,
  scale = "mean",
  remove = FALSE,
  cat = TRUE,
  model.ave = TRUE,
  parallel = FALSE,
  export.package = NULL,
  cores = ifelse(parallel, parallel::detectCores(), 1),
  unsave = FALSE
)
}
\arguments{
\item{posi.model}{glm object for positive catch rate (lognormal or gamma distribution can be used)}

\item{criteria}{モデル選択の基準（他にBIC/AICcが選択可）}

\item{max.narea}{maximum number of areas}

\item{trend}{年トレンドを計算するか}

\item{delta}{delta AIC (or BIC/AICc) がこの値よりも大きくなるまで海区分けを実行}

\item{factor}{年トレンドを算出する際にexpand.gridをするfactorial変数　Yearを最初にAreaを二番目にする（Area-weighted averageにする場合）}

\item{numeric}{年トレンドを算出する際にexpand.gridをする連続変数}

\item{seq.length}{expand.gridをする連続変数を何個に分けるか}

\item{offset}{offset項に使った変数（"Effort"など）※logはいらない}

\item{area.w}{各海区の重みづけ（FALSEですべて同じ、TRUEで緯度経度で面積重みづけ）}

\item{lon.name}{データセットの経度の列名}

\item{lat.name}{データセットの緯度の列名}

\item{seq.lon}{resolution of longitudinal degree}

\item{seq.lat}{resolution of latidudinal degree}

\item{model.ave}{model averagingするかどうか}

\item{parallel}{TRUEでforeachを使ったparallel計算}

\item{export.package}{parallel 計算するときにexportするpackage("speedglm", "glm2"など)}

\item{cores}{並列計算に使用するコア数}

\item{unsave}{delta-GLMの結果をsaveするか unsave=TRUEとすれば保存しない（メモリの節約になる）}

\item{binomial.model}{glm object with a binomial distribution for positive catch probability}
}
\description{
This function is used to conduct an automatic area stratification based on information criteria (AIC, AICc, or BIC) with delta-GLM
}
