% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaGLMtree.R
\encoding{UTF-8}
\name{update_deltaGLM}
\alias{update_deltaGLM}
\title{Reanalyzing delta-GLM and Calculating Year trend}
\usage{
update_deltaGLM(
  binom.model,
  posi.model,
  trend = FALSE,
  factor = c("Year", "Area"),
  numeric = NULL,
  seq.length = 10,
  offset = NULL,
  area.w = TRUE,
  lon.name = "Lon",
  lat.name = "Lat",
  min.lon = 140,
  max.lon = 185,
  min.lat = 30,
  max.lat = 50,
  seq.lon = 2.5,
  seq.lat = 2.5,
  scale = "mean",
  remove = FALSE,
  data = binom.model$data,
  pull.LL = 0
)
}
\arguments{
\item{posi.model}{glm object for positive catch rate (lognormal or gamma distribution can be used)}

\item{trend}{年トレンドを計算するか}

\item{factor}{年トレンドを算出する際にexpand.gridをするfactorial変数　Yearを最初にAreaを二番目にする（Area-weighted averageにする場合）}

\item{numeric}{年トレンドを算出する際にexpand.gridをする連続変数}

\item{seq.length}{expand.gridをする連続変数を何個に分けるか}

\item{offset}{offset項に使った変数（"Effort"など）※logはいらない}

\item{area.w}{各海区の重みづけ（FALSEですべて同じ、TRUEで緯度経度で面積重みづけ）}

\item{lon.name}{データセットの経度の列名}

\item{lat.name}{データセットの緯度の列名}

\item{seq.lon}{resolution of longitudinal degree}

\item{seq.lat}{resolution of latidudinal degree}

\item{binomial.model}{glm object with a binomial distribution for positive catch probability}
}
\description{
This function is used to reanalyze a delta-GLM by excluding categories whose all samples have CPUE > 0 or CPUE = 0 (because of not being able to use a binomial model) and to obtain a year trend
}
